+++ src/Domain/Model/Type.php
@@ -111,10 +111,10 @@ class Type implements WhoInterface, TypeInterface
     public $canSubmitValidation = true;
 
     /** @var bool */
-    public $displayAnalyticsOnSheet = false;
+    private $displayAnalyticsOnSheet = false;
 
     /** @var bool */
-    public $displayAnalyticsOnMeetingList = false;
+    private $displayAnalyticsOnCatalog = false;
 
     public function __construct(Event $event)
     {
@@ -462,7 +462,7 @@ class Type implements WhoInterface, TypeInterface
         bool $mustEvaluateMeeting = false,
         bool $canSubmitValidation = true,
         bool $displayAnalyticsOnSheet = false,
-        bool $displayAnalyticsOnMeetingList = false
+        bool $displayAnalyticsOnCatalog = false
     ) {
         $this->position = $rank;
         $this->hidden = $hidden;
@@ -481,7 +481,7 @@ class Type implements WhoInterface, TypeInterface
         $this->mustEvaluateMeeting = $mustEvaluateMeeting;
         $this->canSubmitValidation = $canSubmitValidation;
         $this->displayAnalyticsOnSheet = $displayAnalyticsOnSheet;
-        $this->displayAnalyticsOnMeetingList = $displayAnalyticsOnMeetingList;
+        $this->displayAnalyticsOnCatalog = $displayAnalyticsOnCatalog;
     }
 
     public function getNumberOfMeetingsPerPlanning(): ?int
@@ -559,5 +559,15 @@ class Type implements WhoInterface, TypeInterface
     public function canSubmitValidation(): bool
     {
         return $this->canSubmitValidation;
+    }
+
+    public function canDisplayAnalyticsOnSheet(): bool
+    {
+        return $this->displayAnalyticsOnSheet;
+    }
+
+    public function canDisplayAnalyticsOnCatalog(): bool
+    {
+        return $this->displayAnalyticsOnCatalog;
     }
 }