+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2208 - Filter meeting request by sheet viewed
+
 ### Fixed
 - VIMEET-2219 - Apply filter on users added on the fly
 - Add sheet card tag for datetime